refactor and api singleton pattern
